requirements.txt
streamlit
pillow
transformers
qdrant-client
torch
sqlite3